update_organization stripe_customer_id param + .rbenv update (#346)
* Update .ruby-version

Problem:
Upon opening this repo, I tried to run `rbenv install` to install the
ruby version required for this project. I was met with the following
error as Ruby 3.0.2 is EOL.

`BUILD FAILED (macOS 15.2 on arm64 using ruby-build 20250115)`

Solution:

I noticed in one of the vcr tests that the user agent was using Ruby
3.1.4. After updating the .ruby-version to 3.1.4 (Which is what I am
assuming the last engineer used), `rbenv install` worked properly and I
was able to bundle and begin contribution.

* Add stripe_customer_id to update_organization

Adds stripe_customer_id as an optional param to update_organization.
This was added in a similar fashion to the node js implementation
(https://workos.com/docs/reference/organization/update) and to match
the initial entitlements documentation
(https://workos.com/docs/user-management/entitlements/use-the-entitlements-in-your-application).

Consumers should now be able to simply call
WorkOS::Organization.update_organization() with a stripe_customer_id to
setup stipe entitlements with their WorkOS orgs.