Add external_id to user and organization (#381)
* Add external_id to user and organization

On create and update (@mxlje already added to user update).

* Allow passing of nil for external_id

External IDs can actually be set to null. On create this isn't an issue,
not sending it will result in it being null. But on update if a user
passes nil we want to be sure a null value goes to the API.

Since organizations and user update methods are different I opted for a
slightly difference solution in each place.

* Test that external_id can be set to null