Add screen_hint to UserManagement
As a solution to build "Sign-Up" paths in Rails applications, this adds
the `screen_hint` to the UserManagement.authorization_url flow. This
appears to be a valid api in NodeJS already (https://workos.com/docs/reference/authkit/authentication/get-authorization-url),
this simply adds it to the Ruby Gem.

Unit tests have been added as well.