Add screen_hint to Ruby Gem + vcr 6 (#389)
* Add screen_hint to UserManagement

As a solution to build "Sign-Up" paths in Rails applications, this adds
the `screen_hint` to the UserManagement.authorization_url flow. This
appears to be a valid api in NodeJS already (https://workos.com/docs/reference/authkit/authentication/get-authorization-url),
this simply adds it to the Ruby Gem.

Unit tests have been added as well.

* Update VCR to resolve global_hook failure

From research, this is a pretty well known issue in vcr 5 that has been
resolved in vcr 6. Running bundle exec rspec still passes with this
update.

Resolves .rbenv/versions/3.1.4/lib/ruby/gems/3.1.0/gems/vcr-5.0.0/lib/vcr/library_hooks/webmock.rb:36:in
`block in global_hook_disabled_requests': wrong number of arguments
(given 1, expected 0) (ArgumentError).